% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetChan.R
\name{chan_poi}
\alias{chan_poi}
\title{A function estimate the location of change point}
\usage{
chan_poi(data, method, loc)
}
\arguments{
\item{data}{the sample.}

\item{method}{Choose the depth function}

\item{loc}{If loc=1,tell us the absolute location, otherwise return the the relative location of the change point.}
}
\description{
The estimation of the location is the point that makes z take the maximum value.
}
\examples{
\dontrun{
library(mvtnorm)
set.seed(518)
mu1<-rep(1,20)
sig1<-diag(10,20)
sig2<-diag(1:20)
mu2<-seq(from=0.55,to=1.5,by=.05)
x<-rmvnorm(200,mu1,sig2)
y<-rmvnorm(100,mu2,sig2)
data<-rbind(x,y)
chan_poi(data,1,2)
  }
}
