% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetChan.R
\name{data_depth}
\alias{data_depth}
\title{A function computes the depth of each sample point}
\usage{
data_depth(data, method)
}
\arguments{
\item{data}{The sample.}

\item{method}{Choose the depth function}
}
\value{
A vector with dimension equal to the sample size
}
\description{
This function calls the depth function in the depth package.
}
\examples{
\dontrun{
x<-matrix(rnorm(300),nc=3)
data_depth(x,1)
  }
}
