% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetChan.R
\name{depth_rank}
\alias{depth_rank}
\title{A function computes the depth of each sample point}
\usage{
depth_rank(data, method)
}
\arguments{
\item{data}{the sample}

\item{method}{Choose the depth function}
}
\value{
A vector with dimension equal to the sample size
}
\description{
The function computes the rank of each sample point by depth.
}
\examples{
\dontrun{
x<-matrix(rnorm(300),nc=3)
rank_vector<-depth_rank(x)
  }
}
