% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sampleC}
\alias{gibbs_sampleC}
\title{A Gibbs sampler using Rcpp}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
