\name{StatComp21086-package}
\alias{StatComp21086-package}
\alias{StatComp21086}
\docType{package}
\title{
  This package provides several functions for estimating the location of a change point
}
\description{
  This package is to estimate the change point based on a new rank-CUSUM proposed in 2020.
}

\author{
Li Xiangchao

Maintainer: Li Xiangchao <lxc1468401361@mail.ustc.edu.cn>
}

\keyword{ package }