% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetChan.R
\name{Z.t}
\alias{Z.t}
\title{A function calculates the value of the statistic Z at t}
\usage{
Z.t(data, method, t)
}
\arguments{
\item{data}{the sample.}

\item{method}{Choose the depth function}

\item{t}{a constant which is between 0 and 1, denotes the location of the data,for example, size of a sample is 100,t=.2 ,that is the 20th point of the sample}
}
\value{
A scalar.
}
\description{
Z is the rank-CUSUM statistic.
}
\examples{
\dontrun{
x<-matrix(rnorm(300),nc=6)
y<-matrix(runif(300),nc=6)
data<-rbind(x,y)
t=.2
Z.t(data,1,t)
  }
}
