% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetChan.R
\name{Z_stat}
\alias{Z_stat}
\title{A function computes the value of Z for all sample points}
\usage{
Z_stat(data, method)
}
\arguments{
\item{data}{the sample}

\item{method}{Choose the depth function}
}
\description{
Z is the rank-CUSUM statistic.
}
\examples{
\dontrun{
x<-matrix(rnorm(300,0,1),nc=6)
y<-matrix(runif(300,1,2),nc=6)
data<-rbind(x,y)
z<-Z_stat(data,1)
plot(z,type="l",col="blue")

data1<-matrix(rnorm(30000),nc=30)
z1<-Z_stat(data1,1)
plot(z1,type="l",col="blue")
  }
}
